requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
discord.py==2.1.0
exceptiongroup==1.0.4
frozenlist==1.3.3
idna==3.4
iniconfig==1.1.1
multidict==6.0.3
packaging==22.0
Pillow==9.3.0
pluggy==1.0.0
pydantic==1.10.2
pytest==7.2.0
pytest-asyncio==0.20.3
python-dotenv==0.21.0
sentry-sdk==1.11.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
